refactor: extract App/ namespace again
This time without having to rewrite all the namespaces as this is being
handled by setting useAppPath and the correct autoload settings in our
composer.json. Still cant use Artisan to make files under the Domain/,
Support/ and Infrastructure/ namespaces but that's a minor issue I'm
going to park for now.