feat(traefik): docker and kubernetes ready reverse proxy
Kamal uses Traefik as a reverse proxy that automatically picks up on new services and routes. In order to work with Kamal effectively I had to learn how to work with Traefik and figure out how it could work together with NGINX and Caddy. As it stands the deploy.yml still contains outdated configuration which I will be tackling next.

I've configured Traefik to either handle generating the SSL certificates using Let's Encrypt and a Cloudflare DNS challenge or pass along the TLS request unprocessed to services that handle certificate generation themselves using tcp.tls.passthrough. As it stands, Caddy is configured to use a self signed certificate since it will be running on the same server. If at some point in time you wish to host these services on a separate box, there are some changes to be made in the Caddyfile.