fix: workaround unstable healthcheck attempts
The healthcheck sometimes fails, possibly because setup without traefik takes less time and the already running load balancer picks up on the newly added service with a short delay. We increase the amount of attempts to temporarily work around this problem.

It should be possible to configure the interval time but that does not appear to work at the moment. I've included it in case that gets resolved. Deployments will take a little longer than they have to but this two minute grace period should result in more stable deploys over all.